+++ Haimen/GUI/DevCompanyList.cs
@@ -75,9 +75,15 @@ namespace Haimen.GUI
             if (gridView1.FocusedRowHandle < 0)
                 return;
 
+            long id = long.Parse(gridView1.GetRowCellValue(gridView1.FocusedRowHandle, col_id).ToString());
+            Company c = Company.CreateByID(id);
+            if (!c.CanDelete(id))
+            {
+                MessageBox.Show(c.ErrorString, "注意");
+                return;
+            }
             if (MessageBox.Show(this, "是否要删除指定的单位？", "警告", MessageBoxButtons.YesNoCancel) == DialogResult.Yes)
             {
-                long id = long.Parse(gridView1.GetRowCellValue(gridView1.FocusedRowHandle, col_id).ToString());
                 Company.Delete(id);
                 gridView1.DeleteRow(gridView1.FocusedRowHandle);
             }
